@@ -19,7 +19,7 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 version: 1.0.0+1
 
 environment:
-  sdk: '>=3.7.2 <4.0.0'
+  sdk: ^3.8.1
 
 # Dependencies specify other packages that your package needs in order to work.
 # To automatically upgrade your package dependencies to the latest versions
@@ -29,6 +29,8 @@ environment:
 # versions available, run `flutter pub outdated`.
 dependencies:
   flutter:
+    sdk: flutter
+  flutter_localizations:
     sdk: flutter
   flutter_web_plugins:
     sdk: flutter
@@ -69,10 +71,30 @@ dependencies:
   firebase_messaging: ^16.0.0
   firebase_storage: ^13.0.0
   firebase_database: ^12.0.0
+  dio: ^5.9.0
+  flutter_bloc: ^9.1.1
+  cached_network_image: ^3.4.1
+  hive_flutter: ^1.1.0
+  path_provider: ^2.1.5
+  injectable: ^2.5.1
+  logger: ^2.6.1
+  flutter_screenutil: ^5.9.3
+  permission_handler: ^12.0.1
+  network_logger: ^1.0.4
+  internet_connection_checker_plus: ^2.7.2
+  intl: ^0.20.2
+  bot_toast: ^4.1.3
+  device_info_plus: ^11.5.0
+  package_info_plus: ^8.3.1
+  flutter_secure_storage: ^9.2.4
 
 dev_dependencies:
   flutter_test:
     sdk: flutter
+  build_runner:
+  intl_utils: ^2.8.3
+  hive_generator:
+  injectable_generator:
 
   # The "flutter_lints" package below contains a set of recommended lints to
   # encourage good coding practices. The lint set provided by the package is
@@ -87,6 +109,8 @@ dev_dependencies:
 # The following section is specific to Flutter packages.
 flutter:
 
+  generate: true
+
   # The following line ensures that the Material Icons font is
   # included with your application, so that you can use the icons in
   # the material Icons class.
@@ -98,28 +122,3 @@ flutter:
     - assets/images/
     - assets/images/projects/
 
-  # An image asset can refer to one or more resolution-specific "variants", see
-  # https://flutter.dev/to/resolution-aware-images
-
-  # For details regarding adding assets from package dependencies, see
-  # https://flutter.dev/to/asset-from-package
-
-  # To add custom fonts to your application, add a fonts section here,
-  # in this "flutter" section. Each entry in this list should have a
-  # "family" key with the font family name, and a "fonts" key with a
-  # list giving the asset and other descriptors for the font. For
-  # example:
-  # fonts:
-  #   - family: Schyler
-  #     fonts:
-  #       - asset: fonts/Schyler-Regular.ttf
-  #       - asset: fonts/Schyler-Italic.ttf
-  #         style: italic
-  #   - family: Trajan Pro
-  #     fonts:
-  #       - asset: fonts/TrajanPro.ttf
-  #       - asset: fonts/TrajanPro_Bold.ttf
-  #         weight: 700
-  #
-  # For details regarding fonts from package dependencies,
-  # see https://flutter.dev/to/font-from-package
